requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
weaviate-client>=3.25.0
openai>=1.0.0
sentence-transformers>=2.2.0
langchain-core>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.1.0
tiktoken>=0.5.0
